Movie_database_management_system/requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
python-dotenv==0.19.0